add error message for inputs
